+++ dimas/src/com/message.rs
@@ -1,6 +1,7 @@
 // Copyright © 2024 Stephan Kunz
 
 // region:		--- modules
+use crate::prelude::{Encode, encode};
 use std::ops::Deref;
 use zenoh::{prelude::sync::SyncResolve, queryable::Query, sample::Sample};
 // endregion:	--- modules
@@ -40,10 +41,10 @@ impl Request {
 	///
 	pub fn reply<T>(&self, value: T)
 	where
-		T: bitcode::Encode,
+		T: Encode,
 	{
 		let key = self.query.selector().key_expr.to_string();
-		let encoded: Vec<u8> = bitcode::encode(&value).expect("should never happen");
+		let encoded: Vec<u8> = encode(&value).expect("should never happen");
 		let sample = Sample::try_from(key, encoded).expect("should never happen");
 
 		self.query