+++ examples/liveliness/src/liveliness.rs
@@ -39,7 +39,7 @@ fn delete_subscription(
 #[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
-	tracing_subscriber::fmt().init();
+	tracing_subscriber::fmt::init();
 
 	// parse arguments
 	let args = Args::parse();