+++ examples/pingpong/src/pong.rs
@@ -2,7 +2,6 @@
 //! Copyright © 2024 Stephan Kunz
 
 // region:		--- modules
-use bitcode::{Decode, Encode};
 use chrono::Local;
 use clap::Parser;
 use dimas::prelude::*;
@@ -34,7 +33,7 @@ fn ping_received(
 	_props: &Arc<RwLock<AgentProps>>,
 	message: &Message,
 ) {
-	let mut message: PingPongMessage = bitcode::decode(message).expect("should not happen");
+	let mut message: PingPongMessage = decode(message).expect("should not happen");
 
 	// set receive-timestamp
 	message.received = Local::now().naive_utc().timestamp_nanos_opt();
@@ -50,7 +49,7 @@ fn ping_received(
 #[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
-	tracing_subscriber::fmt().init();
+	tracing_subscriber::fmt::init();
 
 	// parse arguments
 	let args = Args::parse();