+++ examples/pubsub/src/publisher.rs
@@ -26,7 +26,7 @@ struct AgentProps {
 #[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
-	tracing_subscriber::fmt().init();
+	tracing_subscriber::fmt::init();
 
 	// parse arguments
 	let args = Args::parse();