+++ examples/queries/src/query.rs
@@ -26,14 +26,14 @@ fn query_callback(
 	_props: &Arc<RwLock<AgentProps>>,
 	response: &Message,
 ) {
-	let message: String = bitcode::decode(response).expect("should not happen");
+	let message: String = decode(response).expect("should not happen");
 	println!("Response '{}'", &message);
 }
 
 #[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
-	tracing_subscriber::fmt().init();
+	tracing_subscriber::fmt::init();
 
 	// parse arguments
 	let args = Args::parse();